requirements.txt
django==5.2.6
djangorestframework==3.16.1
pytest-django==4.11.1
Faker==38.2.0